+++ src/Classes/ImportBuchungen.php
@@ -68,7 +68,7 @@ class ImportBuchungen extends Backend
 
 			foreach ($arrUploaded as $txtFile)
 			{
-				$objFile = new File($txtFile, true);
+				$objFile = new File($txtFile);
 
 				if ($objFile->extension != 'csv')
 				{